Additional zoom settings.
Add settings for PDF recache zoom difference threshold. Also adds a
threshold for which zoom snaps to the initial zoom if the difference in
distances between the two touch points is less than some given
percentage of the original.

The goal of this feature is to make it easier to pan with two fingers in
large PDFs/other documents in which zooming is slow.